Move verify function to Verifier
